# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-AI Coding Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: path
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileNode"
  /projects/{id}/files/read:
    post:
      operationId: readFile
      tags: [files]
      summary: Read a file's content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadFileBody"
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
  /projects/{id}/files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write content to a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteFileBody"
      responses:
        "200":
          description: File written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
...
  /memory:
    get:
      operationId: getMemory
      tags: [memory]
      summary: Get agent memory for a project
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
    post:
      operationId: addMemory
      tags: [memory]
      summary: Add a memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemoryBody"
      responses:
        "201":
          description: Memory added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /memory/{id}:
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
        language:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - path
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
      required:
        - name
        - path
    FileNode:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        size:
          type: ["integer", "null"]
        children:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/FileNode"
      required:
        - name
        - path
        - type
    ReadFileBody:
      type: object
      properties:
        path:
          type: string
      required:
        - path
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: ["string", "null"]
        size:
          type: integer
      required:
        - path
        - content
        - size
    WriteFileBody:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        createDirs:
          type: boolean
      required:
        - path
        - content
...
    OperationResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    MemoryEntry:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [project_structure, known_bug, past_fix, important_file, error_pattern, general]
        key:
          type: string
        value:
          type: string
        createdAt:
          type: string
      required:
        - id
        - type
        - key
        - value
        - createdAt
    AddMemoryBody:
      type: object
      properties:
        projectId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [project_structure, known_bug, past_fix, important_file, error_pattern, general]
        key:
          type: string
        value:
          type: string
      required:
        - type
        - key
        - value
...