# lib/api-zod/src/generated/types/fileContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multi-AI Coding Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface FileContent {
  path: string;
  content: string;
  /** @nullable */
  language?: string | null;
  size: number;
}